# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NyayaLabel AI API",
    "version": "1.0.0",
    "description": "API documentation for the NyayaLabel AI backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate Officer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful authentication" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh JWT Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token refreshed successfully" },
          "401": { "description": "Invalid token" }
        }
      }
    },
    "/inspections/sync": {
      "post": {
        "summary": "Sync an inspection",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inspectionId": { "type": "string" },
                  "shopName": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Synced successfully" }
        }
      }
    },
    "/images/upload": {
      "post": {
        "summary": "Upload evidence image",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Image uploaded successfully" }
        }
      }
    },
    "/rules": {
      "get": {
        "summary": "Get all active rules",
        "responses": {
          "200": { "description": "List of active rules" }
        }
      },
      "post": {
        "summary": "Create or update a rule version",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "201": { "description": "Rule created" }
        }
      }
    }
  }
}